# openapi.yaml
openapi: 3.1.0
info:
  title: rsdecider
  version: 0.1.0
  description: |
    Concurrent inference server for Laya decision models. Each question about a `state` is answered
    by the English or multilingual model (Laya routing mode), with micro-batching, a two-tier cache,
    per-key rate limits and load shedding.

    Pipeline order: auth (401) → body limit (413) → JSON parse (422) → per-key rate limit (429) →
    Idempotency-Key (409 / replay / 422) → validation + routing (422) → cache → tokenize misses →
    admission (529) → inference. Any request that
    passes `server.request_timeout_ms` returns 504.
  license:
    name: see repository
servers:
  - url: /
    description: The server hosting this spec (Swagger UI at /docs)
security:
  - apiKey: []

paths:
  /v1/decide:
    post:
      summary: Answer every question about one state
      operationId: decide
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DecideRequest' }
            example:
              state: Customer was charged twice for invoice #4411 and is asking for a refund.
              questions:
                action:
                  type: choice
                  instructions: How should support handle this?
                  criteria: { approve: issue the refund, deny: refuse the refund, escalate: send to a human agent }
                urgent: { type: noul, instructions: Is this urgent? }
                severity: { type: score, instructions: How severe is the issue?, criteria: [low, medium, high] }
      responses:
        '200':
          description: Answers, keyed by question id. The id is also in `x-request-id`.
          headers:
            x-request-id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    required: [id]
                    properties:
                      id: { type: string, example: req_01M31BAVX63WFSFXX0WQWM0FAK }
                  - $ref: '#/components/schemas/DecideResult'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/InProgress' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '422': { $ref: '#/components/responses/Invalid' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
        '504': { $ref: '#/components/responses/Deadline' }
        '529': { $ref: '#/components/responses/Overloaded' }

  /v1/decide/batch:
    post:
      summary: Answer several states in one call (1..limits.max_batch_states)
      description: Each item is routed on its own. One invalid item fails the whole batch (422 names `items[i]`). Admission is all-or-nothing, so a batch gets 529 unless all of its questions fit.
      operationId: decideBatch
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  minItems: 1
                  items: { $ref: '#/components/schemas/DecideRequest' }
      responses:
        '200':
          description: One result per item, in request order.
          headers:
            x-request-id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema:
                type: object
                required: [id, results]
                properties:
                  id: { type: string }
                  results:
                    type: array
                    items: { $ref: '#/components/schemas/DecideResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/InProgress' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '422': { $ref: '#/components/responses/Invalid' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
        '504': { $ref: '#/components/responses/Deadline' }
        '529': { $ref: '#/components/responses/Overloaded' }

  /healthz:
    get:
      summary: Liveness
      operationId: healthz
      security: []
      responses:
        '200': { description: Process is up, content: { text/plain: { schema: { type: string, const: ok } } } }

  /openapi.yaml:
    get:
      summary: This OpenAPI document
      operationId: openapi
      security: []
      responses:
        '200': { description: OpenAPI 3.1 YAML, content: { application/yaml: { schema: { type: string } } } }

  /docs:
    get:
      summary: Swagger UI for this API
      operationId: docs
      security: []
      responses:
        '200': { description: HTML page, content: { text/html: { schema: { type: string } } } }

  /readyz:
    get:
      summary: Readiness — every model has at least one live worker
      operationId: readyz
      security: []
      responses:
        '200': { description: Ready, content: { text/plain: { schema: { type: string, const: ready } } } }
        '503': { description: A model has no live workers, content: { text/plain: { schema: { type: string, const: not ready } } } }

components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: Keys come from `[[keys]]` in rsdecider.toml (only SHA-256 hashes are stored; `rsdecider hash-key <secret>`).

  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Visible ASCII. A repeat with the same body replays the stored 200 response (same id); while the first is running → 409; a different body → 422. Errors are never stored.
      schema: { type: string }

  headers:
    RequestId:
      schema: { type: string }
    RetryAfter:
      description: Seconds to wait before retrying.
      schema: { type: integer, minimum: 1 }

  schemas:
    DecideRequest:
      type: object
      required: [state, questions]
      properties:
        state:
          description: The situation to decide about — a string or any JSON value (at most limits.max_state_chars).
        model:
          type: string
          description: Force a model by name (routing reason becomes `explicit`). If omitted, the language decides.
        questions:
          type: object
          description: Question id → question (at most limits.max_questions_per_state).
          additionalProperties: { $ref: '#/components/schemas/Question' }

    Question:
      type: object
      required: [type, instructions]
      properties:
        type: { type: string, enum: [choice, score, noul] }
        instructions: { description: Question text — a string or any JSON value. }
        criteria:
          description: |
            choice/score: options as a list (`["low","high"]`) or a label → description map.
            noul (yes/no): optional map, e.g. `{"true": "..."}`.
          oneOf:
            - type: array
              items: { type: string }
            - type: object
              additionalProperties: { type: string }

    DecideResult:
      type: object
      required: [model, answers, usage, routing, timing_ms]
      properties:
        model: { type: string, example: laya-english@c3e0839d173a }
        answers:
          type: object
          additionalProperties: { $ref: '#/components/schemas/Answer' }
        usage:
          type: object
          properties:
            input_tokens: { type: integer }
            output_tokens: { type: integer, const: 0 }
        routing:
          type: object
          properties:
            model: { type: string, example: english }
            reason: { type: string, enum: [detected:english, detected:non_english, explicit] }
        timing_ms:
          type: object
          properties:
            tokenize: { type: number }
            wait: { type: number }
            total: { type: number }

    Answer:
      type: object
      required: [type, confidence, action, cached]
      properties:
        type: { type: string, enum: [choice, score, noul] }
        choice: { type: string, description: choice only — the most probable label }
        score: { type: number, description: score only — expected option index }
        legend: { type: object, additionalProperties: { type: string }, description: score only — index → label }
        noul: { type: number, description: noul only — probability of yes }
        probabilities: { type: object, additionalProperties: { type: number }, description: choice/score — per option }
        confidence: { type: number }
        action:
          type: object
          properties:
            act_probability: { type: number }
        cached: { type: boolean }
      example:
        type: choice
        choice: escalate
        probabilities: { approve: 0.4047, deny: 0.1535, escalate: 0.4417 }
        confidence: 0.0764
        action: { act_probability: 1.0 }
        cached: false

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [unauthorized, idempotency_in_progress, payload_too_large, invalid_request, rate_limited, overloaded, deadline_exceeded, internal]
            message: { type: string }

  responses:
    Unauthorized:
      description: Missing or unknown API key
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    InProgress:
      description: A request with this Idempotency-Key is still running
      headers: { Retry-After: { $ref: '#/components/headers/RetryAfter' } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    TooLarge:
      description: Body exceeds limits.max_body_bytes
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Invalid:
      description: Invalid JSON or field (message names the path, e.g. `questions.q` or `items[1].`), unknown model, bad Idempotency-Key, or a reused key with a different body
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    RateLimited:
      description: Per-key rate limit exceeded (a batch costs one token per item)
      headers: { Retry-After: { $ref: '#/components/headers/RetryAfter' } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Internal:
      description: Inference failed (for example a worker panic); the next request is served normally
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Deadline:
      description: The request passed server.request_timeout_ms
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Overloaded:
      description: The model queue cannot take these questions within the deadline; retry later
      headers: { Retry-After: { $ref: '#/components/headers/RetryAfter' } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }